1744 - Reorganize, SF
