+++ server/ee/apps/worker-app/build.gradle.kts
@@ -50,6 +50,7 @@ dependencies {
     implementation(project(":server:libs:modules:components:accelo"))
     implementation(project(":server:libs:modules:components:active-campaign"))
     implementation(project(":server:libs:modules:components:affinity"))
+    implementation(project(":server:libs:modules:components:ai:ai-text-analysis"))
     implementation(project(":server:libs:modules:components:airtable"))
     implementation(project(":server:libs:modules:components:aitable"))
     implementation(project(":server:libs:modules:components:app-event"))
@@ -95,7 +96,6 @@ dependencies {
     implementation(project(":server:libs:modules:components:json-file"))
     implementation(project(":server:libs:modules:components:json-helper"))
     implementation(project(":server:libs:modules:components:keap"))
-    implementation(project(":server:libs:modules:components:llm:ai-text-analysis"))
     implementation(project(":server:libs:modules:components:llm:amazon-bedrock"))
     implementation(project(":server:libs:modules:components:llm:anthropic"))
     implementation(project(":server:libs:modules:components:llm:azure-openai"))
