+++ server/libs/modules/components/ai/ai-text-analysis/src/main/java/com/bytechef/component/ai/text/analysis/constant/AiTextAnalysisConstants.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2023-present ByteChef Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.component.ai.text.analysis.constant;
+/**
+ * @author Marko Krišković
+ */
+public class AiTextAnalysisConstants {
+
+    public static final String TEXT = "text";
+    public static final String FORMAT = "format";
+    public static final String PROMPT = "prompt";
+    public static final String MODEL_PROVIDER = "modelProvider";
+    public static final String SUMMARIZE_TEXT = "summarizeText";
+
+    private AiTextAnalysisConstants() {
+    }
+}
