+++ server/libs/modules/components/code-workflow/src/main/java/com/bytechef/component/codeworkflow/action/CodeWorkflowPerformAction.java
@@ -21,17 +21,8 @@ import static com.bytechef.component.definition.ComponentDsl.action;
 import static com.bytechef.component.definition.ComponentDsl.integer;
 import static com.bytechef.component.definition.ComponentDsl.string;
 
+import com.bytechef.component.codeworkflow.action.definition.CodeWorkflowPerformActionDefinition;
 import com.bytechef.component.codeworkflow.task.CodeWorkflowTaskExecutor;
-import com.bytechef.component.definition.ActionContext;
-import com.bytechef.component.definition.ActionDefinition;
-import com.bytechef.component.definition.Parameters;
-import com.bytechef.platform.component.constant.MetadataConstants;
-import com.bytechef.platform.component.definition.AbstractActionDefinitionWrapper;
-import com.bytechef.platform.component.definition.MultipleConnectionsPerformFunction;
-import com.bytechef.platform.component.definition.ParameterConnection;
-import com.bytechef.platform.constant.ModeType;
-import java.util.Map;
-import java.util.Optional;
 
 /**
  * @author Ivica Cardic
@@ -53,32 +44,5 @@ public class CodeWorkflowPerformAction {
                     string("taskName")
                         .label("Task Name")),
             codeWorkflowTaskExecutor);
-    }
-
-    public static class CodeWorkflowPerformActionDefinition extends AbstractActionDefinitionWrapper {
-
-        private final CodeWorkflowTaskExecutor codeWorkflowTaskExecutor;
-
-        public CodeWorkflowPerformActionDefinition(ActionDefinition actionDefinition,
-            CodeWorkflowTaskExecutor codeWorkflowTaskExecutor) {
-            super(actionDefinition);
-
-            this.codeWorkflowTaskExecutor = codeWorkflowTaskExecutor;
-        }
-
-        @Override
-        public Optional<PerformFunction> getPerform() {
-            return Optional.of((MultipleConnectionsPerformFunction) this::perform);
-        }
-
-        protected Object perform(
-            Parameters inputParameters, Map<String, ? extends ParameterConnection> connectionParameters,
-            Parameters extensions, ActionContext actionContext) {
-
-            return codeWorkflowTaskExecutor.executePerform(
-                inputParameters.getRequiredString("codeWorkflowContainerReference"),
-                inputParameters.getRequiredString("workflowName"), inputParameters.getRequiredString("taskName"),
-                inputParameters.getRequired(MetadataConstants.TYPE, ModeType.class));
-        }
     }
 }
