--- server/libs/modules/components/llm/ai-text-analysis/src/main/java/com/bytechef/component/ai/text/analysis/constant/AiTextAnalysisConstants.java
@@ -1,12 +1,0 @@
-package com.bytechef.component.ai.text.analysis.constant;
-/**
- * @author Marko Krišković
- */
-public class AiTextAnalysisConstants {
-    public static final String TEXT = "text";
-    public static final String FORMAT = "format";
-    public static final String PROMPT = "prompt";
-    public static final String CONNECTION_PROVIDER = "connectionProvider";
-
-    private AiTextAnalysisConstants() {}
-}
