+++ server/libs/modules/components/llm/amazon-bedrock/src/main/java/com/bytechef/component/amazon/bedrock/action/AmazonBedrockAnthropic2ChatAction.java
@@ -36,18 +36,13 @@ import static com.bytechef.component.llm.constant.LLMConstants.TOP_K_PROPERTY;
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P;
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P_PROPERTY;
 
+import com.bytechef.component.amazon.bedrock.constant.AmazonBedrockConstants;
 import com.bytechef.component.definition.ActionContext;
-import com.bytechef.component.definition.ComponentDsl;
 import com.bytechef.component.definition.ComponentDsl.ModifiableActionDefinition;
-import com.bytechef.component.definition.Option;
 import com.bytechef.component.definition.Parameters;
 import com.bytechef.component.definition.TypeReference;
 import com.bytechef.component.llm.Chat;
-import com.bytechef.component.llm.util.LLMUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.ai.bedrock.anthropic.AnthropicChatOptions;
 import org.springframework.ai.bedrock.anthropic.BedrockAnthropicChatModel;
 import org.springframework.ai.bedrock.anthropic.api.AnthropicChatBedrockApi;
@@ -60,13 +55,6 @@ import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
  */
 public class AmazonBedrockAnthropic2ChatAction {
 
-    public static final List<Option<String>> MODELS_ENUM = LLMUtils
-        .getEnumOptions(Arrays.stream(AnthropicChatBedrockApi.AnthropicChatModel.values())
-        .collect(
-            Collectors.toMap(
-                AnthropicChatBedrockApi.AnthropicChatModel::getName,
-                AnthropicChatBedrockApi.AnthropicChatModel::getName, (f, s) -> f)));
-
     public static final ModifiableActionDefinition ACTION_DEFINITION = action("askAnthropic2")
         .title("Ask Anthropic2")
         .description("Ask anything you want.")
@@ -75,7 +63,7 @@ public class AmazonBedrockAnthropic2ChatAction {
                 .label("Model")
                 .description("ID of the model to use.")
                 .required(true)
-                .options(MODELS_ENUM),
+                .options(AmazonBedrockConstants.ANTHROPIC2_MODELS),
             MESSAGES_PROPERTY,
             integer(MAX_TOKENS)
                 .label("Max Tokens")
@@ -100,16 +88,6 @@ public class AmazonBedrockAnthropic2ChatAction {
     }
 
     public static final Chat CHAT = new Chat() {
-        @Override
-        public ChatOptions createChatOptions(Parameters inputParameters) {
-            return AnthropicChatOptions.builder()
-                .withTemperature(inputParameters.getDouble(TEMPERATURE))
-                .withMaxTokensToSample(inputParameters.getInteger(MAX_TOKENS))
-                .withTopP(inputParameters.getDouble(TOP_P))
-                .withStopSequences(inputParameters.getList(STOP, new TypeReference<>() {}))
-                .withTopK(inputParameters.getInteger(TOP_K))
-                .build();
-        }
 
         @Override
         public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
@@ -121,6 +99,16 @@ public class AmazonBedrockAnthropic2ChatAction {
                         connectionParameters.getRequiredString(SECRET_ACCESS_KEY)),
                     connectionParameters.getRequiredString(REGION), new ObjectMapper()),
                 (AnthropicChatOptions) createChatOptions(inputParameters));
+        }
+
+        private ChatOptions createChatOptions(Parameters inputParameters) {
+            return AnthropicChatOptions.builder()
+                .withTemperature(inputParameters.getDouble(TEMPERATURE))
+                .withMaxTokensToSample(inputParameters.getInteger(MAX_TOKENS))
+                .withTopP(inputParameters.getDouble(TOP_P))
+                .withStopSequences(inputParameters.getList(STOP, new TypeReference<>() {}))
+                .withTopK(inputParameters.getInteger(TOP_K))
+                .build();
         }
     };
 }
