+++ server/libs/modules/components/llm/amazon-bedrock/src/main/java/com/bytechef/component/amazon/bedrock/action/AmazonBedrockCohereChatAction.java
@@ -45,16 +45,15 @@ import static com.bytechef.component.llm.constant.LLMConstants.TOP_K_PROPERTY;
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P;
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P_PROPERTY;
 
+import com.bytechef.component.amazon.bedrock.constant.AmazonBedrockConstants;
 import com.bytechef.component.definition.ActionContext;
 import com.bytechef.component.definition.ComponentDsl.ModifiableActionDefinition;
-import com.bytechef.component.definition.Option;
 import com.bytechef.component.definition.Parameters;
 import com.bytechef.component.definition.TypeReference;
 import com.bytechef.component.llm.Chat;
 import com.bytechef.component.llm.util.LLMUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Arrays;
-import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.ai.bedrock.cohere.BedrockCohereChatModel;
 import org.springframework.ai.bedrock.cohere.BedrockCohereChatOptions;
@@ -68,13 +67,6 @@ import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
  */
 public class AmazonBedrockCohereChatAction {
 
-    public static final List<Option<String>> MODELS_ENUM = LLMUtils
-        .getEnumOptions(Arrays.stream(CohereChatBedrockApi.CohereChatModel.values())
-            .collect(
-                Collectors.toMap(
-                    CohereChatBedrockApi.CohereChatModel::getName,
-                    CohereChatBedrockApi.CohereChatModel::getName, (f, s) -> f)));
-
     public static final ModifiableActionDefinition ACTION_DEFINITION = action("askCohere")
         .title("Ask Cohere")
         .description("Ask anything you want.")
@@ -83,7 +75,7 @@ public class AmazonBedrockCohereChatAction {
                 .label("Model")
                 .description("ID of the model to use.")
                 .required(true)
-                .options(MODELS_ENUM),
+                .options(AmazonBedrockConstants.COHERE_MODELS),
             MESSAGES_PROPERTY,
             RESPONSE_FORMAT_PROPERTY,
             RESPONSE_SCHEMA_PROPERTY,
@@ -138,8 +130,20 @@ public class AmazonBedrockCohereChatAction {
     }
 
     public static final Chat CHAT = new Chat() {
+
         @Override
-        public ChatOptions createChatOptions(Parameters inputParameters) {
+        public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
+            return new BedrockCohereChatModel(
+                new CohereChatBedrockApi(
+                    inputParameters.getRequiredString(MODEL),
+                    () -> AwsBasicCredentials.create(
+                        connectionParameters.getRequiredString(ACCESS_KEY_ID),
+                        connectionParameters.getRequiredString(SECRET_ACCESS_KEY)),
+                    connectionParameters.getRequiredString(REGION), new ObjectMapper()),
+                (BedrockCohereChatOptions) createChatOptions(inputParameters));
+        }
+
+        private ChatOptions createChatOptions(Parameters inputParameters) {
             return BedrockCohereChatOptions.builder()
                 .withTemperature(inputParameters.getDouble(TEMPERATURE))
                 .withMaxTokens(inputParameters.getInteger(MAX_TOKENS))
@@ -154,18 +158,6 @@ public class AmazonBedrockCohereChatAction {
                 .withTruncate(inputParameters.get(TRUNCATE,
                     CohereChatBedrockApi.CohereChatRequest.Truncate.class))
                 .build();
-        }
-
-        @Override
-        public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
-            return new BedrockCohereChatModel(
-                new CohereChatBedrockApi(
-                    inputParameters.getRequiredString(MODEL),
-                    () -> AwsBasicCredentials.create(
-                        connectionParameters.getRequiredString(ACCESS_KEY_ID),
-                        connectionParameters.getRequiredString(SECRET_ACCESS_KEY)),
-                    connectionParameters.getRequiredString(REGION), new ObjectMapper()),
-                (BedrockCohereChatOptions) createChatOptions(inputParameters));
         }
     };
 }
