+++ server/libs/modules/components/llm/amazon-bedrock/src/main/java/com/bytechef/component/amazon/bedrock/action/AmazonBedrockLlamaChatAction.java
@@ -32,16 +32,12 @@ import static com.bytechef.component.llm.constant.LLMConstants.TEMPERATURE_PROPE
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P;
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P_PROPERTY;
 
+import com.bytechef.component.amazon.bedrock.constant.AmazonBedrockConstants;
 import com.bytechef.component.definition.ActionContext;
 import com.bytechef.component.definition.ComponentDsl.ModifiableActionDefinition;
-import com.bytechef.component.definition.Option;
 import com.bytechef.component.definition.Parameters;
 import com.bytechef.component.llm.Chat;
-import com.bytechef.component.llm.util.LLMUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.ai.bedrock.llama.BedrockLlamaChatModel;
 import org.springframework.ai.bedrock.llama.BedrockLlamaChatOptions;
 import org.springframework.ai.bedrock.llama.api.LlamaChatBedrockApi;
@@ -54,14 +50,6 @@ import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
  */
 public class AmazonBedrockLlamaChatAction {
 
-    public static final List<Option<String>> MODELS_ENUM = LLMUtils
-        .getEnumOptions(Arrays.stream(LlamaChatBedrockApi.LlamaChatModel.values())
-            .collect(
-                Collectors.toMap(
-                    LlamaChatBedrockApi.LlamaChatModel::getName,
-                    LlamaChatBedrockApi.LlamaChatModel::getName,
-                    (f, s) -> f)));
-
     public static final ModifiableActionDefinition ACTION_DEFINITION = action("askLlama")
         .title("Ask Llama")
         .description("Ask anything you want.")
@@ -70,7 +58,7 @@ public class AmazonBedrockLlamaChatAction {
                 .label("Model")
                 .description("ID of the model to use.")
                 .required(true)
-                .options(MODELS_ENUM),
+                .options(AmazonBedrockConstants.LLAMA_MODELS),
             MESSAGES_PROPERTY,
             RESPONSE_FORMAT_PROPERTY,
             RESPONSE_SCHEMA_PROPERTY,
@@ -90,14 +78,6 @@ public class AmazonBedrockLlamaChatAction {
     }
 
     public static final Chat CHAT = new Chat() {
-        @Override
-        public ChatOptions createChatOptions(Parameters inputParameters) {
-            return BedrockLlamaChatOptions.builder()
-                .withTemperature(inputParameters.getDouble(TEMPERATURE))
-                .withMaxGenLen(inputParameters.getInteger(MAX_TOKENS))
-                .withTopP(inputParameters.getDouble(TOP_P))
-                .build();
-        }
 
         @Override
         public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
@@ -109,6 +89,14 @@ public class AmazonBedrockLlamaChatAction {
                         connectionParameters.getRequiredString(SECRET_ACCESS_KEY)),
                     connectionParameters.getRequiredString(REGION), new ObjectMapper()),
                 (BedrockLlamaChatOptions) createChatOptions(inputParameters));
+        }
+
+        private ChatOptions createChatOptions(Parameters inputParameters) {
+            return BedrockLlamaChatOptions.builder()
+                .withTemperature(inputParameters.getDouble(TEMPERATURE))
+                .withMaxGenLen(inputParameters.getInteger(MAX_TOKENS))
+                .withTopP(inputParameters.getDouble(TOP_P))
+                .build();
         }
     };
 }
