+++ server/libs/modules/components/llm/amazon-bedrock/src/main/java/com/bytechef/component/amazon/bedrock/action/AmazonBedrockTitanChatAction.java
@@ -34,17 +34,13 @@ import static com.bytechef.component.llm.constant.LLMConstants.TEMPERATURE_PROPE
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P;
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P_PROPERTY;
 
+import com.bytechef.component.amazon.bedrock.constant.AmazonBedrockConstants;
 import com.bytechef.component.definition.ActionContext;
 import com.bytechef.component.definition.ComponentDsl.ModifiableActionDefinition;
-import com.bytechef.component.definition.Option;
 import com.bytechef.component.definition.Parameters;
 import com.bytechef.component.definition.TypeReference;
 import com.bytechef.component.llm.Chat;
-import com.bytechef.component.llm.util.LLMUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.ai.bedrock.titan.BedrockTitanChatModel;
 import org.springframework.ai.bedrock.titan.BedrockTitanChatOptions;
 import org.springframework.ai.bedrock.titan.api.TitanChatBedrockApi;
@@ -56,13 +52,6 @@ import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
  * @author Marko Kriskovic
  */
 public class AmazonBedrockTitanChatAction {
-    public static final List<Option<String>> MODELS_ENUM = LLMUtils
-        .getEnumOptions(Arrays.stream(TitanChatBedrockApi.TitanChatModel.values())
-            .collect(
-                Collectors.toMap(
-                    TitanChatBedrockApi.TitanChatModel::getName,
-                    TitanChatBedrockApi.TitanChatModel::getName,
-                    (f, s) -> f)));
 
     public static final ModifiableActionDefinition ACTION_DEFINITION = action("askTitan")
         .title("Ask Titan")
@@ -72,7 +61,7 @@ public class AmazonBedrockTitanChatAction {
                 .label("Model")
                 .description("ID of the model to use.")
                 .required(true)
-                .options(MODELS_ENUM),
+                .options(AmazonBedrockConstants.TITAN_MODELS),
             MESSAGES_PROPERTY,
             RESPONSE_FORMAT_PROPERTY,
             RESPONSE_SCHEMA_PROPERTY,
@@ -93,15 +82,6 @@ public class AmazonBedrockTitanChatAction {
     }
 
     public static final Chat CHAT = new Chat() {
-        @Override
-        public ChatOptions createChatOptions(Parameters inputParameters) {
-            return BedrockTitanChatOptions.builder()
-                .withTemperature(inputParameters.getDouble(TEMPERATURE))
-                .withMaxTokenCount(inputParameters.getInteger(MAX_TOKENS))
-                .withTopP(inputParameters.getDouble(TOP_P))
-                .withStopSequences(inputParameters.getList(STOP, new TypeReference<>() {}))
-                .build();
-        }
 
         @Override
         public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
@@ -112,6 +92,15 @@ public class AmazonBedrockTitanChatAction {
                         connectionParameters.getRequiredString(SECRET_ACCESS_KEY)),
                     connectionParameters.getRequiredString(REGION), new ObjectMapper()),
                 (BedrockTitanChatOptions) createChatOptions(inputParameters));
+        }
+
+        private ChatOptions createChatOptions(Parameters inputParameters) {
+            return BedrockTitanChatOptions.builder()
+                .withTemperature(inputParameters.getDouble(TEMPERATURE))
+                .withMaxTokenCount(inputParameters.getInteger(MAX_TOKENS))
+                .withTopP(inputParameters.getDouble(TOP_P))
+                .withStopSequences(inputParameters.getList(STOP, new TypeReference<>() {}))
+                .build();
         }
     };
 }
