+++ server/libs/modules/components/llm/anthropic/src/main/java/com/bytechef/component/anthropic/action/AnthropicChatAction.java
@@ -35,15 +35,12 @@ import static com.bytechef.component.llm.constant.LLMConstants.TOP_K_PROPERTY;
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P;
 import static com.bytechef.component.llm.constant.LLMConstants.TOP_P_PROPERTY;
 
+import com.bytechef.component.anthropic.constant.AnthropicConstants;
 import com.bytechef.component.definition.ActionContext;
 import com.bytechef.component.definition.ComponentDsl.ModifiableActionDefinition;
-import com.bytechef.component.definition.Option;
 import com.bytechef.component.definition.Parameters;
 import com.bytechef.component.definition.TypeReference;
 import com.bytechef.component.llm.Chat;
-import com.bytechef.component.llm.util.LLMUtils;
-import java.util.Arrays;
-import java.util.stream.Collectors;
 import org.springframework.ai.anthropic.AnthropicChatModel;
 import org.springframework.ai.anthropic.AnthropicChatOptions;
 import org.springframework.ai.anthropic.api.AnthropicApi;
@@ -54,11 +51,6 @@ import org.springframework.ai.chat.prompt.ChatOptions;
  * @author Marko Kriskovic
  */
 public class AnthropicChatAction {
-    public static final List<Option<String>> MODELS_ENUM = LLMUtils
-        .getEnumOptions(Arrays.stream(AnthropicApi.ChatModel.values())
-            .collect(
-                Collectors.toMap(
-                    AnthropicApi.ChatModel::getValue, AnthropicApi.ChatModel::getValue, (f, s) -> f)));
 
     public static final ModifiableActionDefinition ACTION_DEFINITION = action(ASK)
         .title("Ask")
@@ -68,7 +60,7 @@ public class AnthropicChatAction {
                 .label("Model")
                 .description("ID of the model to use.")
                 .required(true)
-                .options(MODELS_ENUM),
+                .options(AnthropicConstants.MODELS),
             MESSAGES_PROPERTY,
             integer(MAX_TOKENS)
                 .label("Max Tokens")
@@ -86,7 +78,6 @@ public class AnthropicChatAction {
     private AnthropicChatAction() {
     }
 
-    public static Object perform(
     public static Object perform(Parameters inputParameters, Parameters connectionParameters, ActionContext context) {
         return CHAT.getResponse(inputParameters, connectionParameters, context);
     }
@@ -94,7 +85,13 @@ public class AnthropicChatAction {
     public static final Chat CHAT = new Chat() {
 
         @Override
-        public ChatOptions createChatOptions(Parameters inputParameters) {
+        public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
+            return new AnthropicChatModel(
+                new AnthropicApi(connectionParameters.getString(TOKEN)),
+                (AnthropicChatOptions) createChatOptions(inputParameters));
+        }
+
+        private ChatOptions createChatOptions(Parameters inputParameters) {
             AnthropicChatOptions.Builder builder = AnthropicChatOptions.builder()
                 .withModel(inputParameters.getRequiredString(MODEL))
                 .withTemperature(inputParameters.getDouble(TEMPERATURE))
@@ -103,20 +100,7 @@ public class AnthropicChatAction {
                 .withStopSequences(inputParameters.getList(STOP, new TypeReference<>() {}))
                 .withTopK(inputParameters.getInteger(TOP_K));
 
-            List<String> functions = inputParameters.getList(FUNCTIONS, new TypeReference<>() {});
-
-            if (functions != null) {
-                builder.withFunctions(new HashSet<>(functions));
-            }
-
             return builder.build();
-        }
-
-        @Override
-        public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
-            return new AnthropicChatModel(
-                new AnthropicApi(connectionParameters.getString(TOKEN)),
-                (AnthropicChatOptions) createChatOptions(inputParameters));
         }
     };
 }
