+++ server/libs/modules/components/llm/azure-openai/src/main/java/com/bytechef/component/azure/openai/action/AzureOpenAIChatAction.java
@@ -95,9 +95,19 @@ public class AzureOpenAIChatAction {
         return CHAT.getResponse(inputParameters, connectionParameters, context);
     }
 
-    private static final Chat CHAT = new Chat() {
+    public static final Chat CHAT = new Chat() {
+
         @Override
-        public ChatOptions createChatOptions(Parameters inputParameters) {
+        public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
+            OpenAIClientBuilder openAIClientBuilder = new OpenAIClientBuilder()
+                .credential(new KeyCredential(connectionParameters.getString(TOKEN)))
+                .endpoint(connectionParameters.getString(ENDPOINT));
+
+            return new AzureOpenAiChatModel(
+                openAIClientBuilder, (AzureOpenAiChatOptions) createChatOptions(inputParameters));
+        }
+
+        private ChatOptions createChatOptions(Parameters inputParameters) {
             Integer responseInteger = inputParameters.getInteger(RESPONSE_FORMAT);
             AzureOpenAiResponseFormat format = responseInteger == null || responseInteger < 1
                 ? AzureOpenAiResponseFormat.TEXT : AzureOpenAiResponseFormat.JSON;
@@ -115,23 +125,7 @@ public class AzureOpenAIChatAction {
                 .withUser(inputParameters.getString(USER))
                 .withResponseFormat(format);
 
-            List<String> functions = inputParameters.getList(FUNCTIONS, new TypeReference<>() {});
-
-            if (functions != null) {
-                builder.withFunctions(new HashSet<>(functions));
-            }
-
             return builder.build();
-        }
-
-        @Override
-        public ChatModel createChatModel(Parameters inputParameters, Parameters connectionParameters) {
-            OpenAIClientBuilder openAIClientBuilder = new OpenAIClientBuilder()
-                .credential(new KeyCredential(connectionParameters.getString(TOKEN)))
-                .endpoint(connectionParameters.getString(ENDPOINT));
-
-            return new AzureOpenAiChatModel(
-                openAIClientBuilder, (AzureOpenAiChatOptions) createChatOptions(inputParameters));
         }
     };
 }
