+++ server/libs/modules/components/script/src/main/java/com/bytechef/component/script/action/ScriptJavaScriptAction.java
@@ -31,7 +31,7 @@ import static com.bytechef.component.script.constant.ScriptConstants.INPUT;
 import static com.bytechef.platform.component.definition.ScriptComponentDefinition.SCRIPT;
 
 import com.bytechef.component.definition.Property;
-import com.bytechef.component.script.definition.ScriptActionDefinition;
+import com.bytechef.component.script.action.definition.ScriptActionDefinition;
 import com.bytechef.component.script.engine.PolyglotEngine;
 
 /**
