+++ settings.gradle.kts
@@ -219,6 +219,7 @@ include("server:libs:platform:platform-workflow:platform-workflow-execution:plat
 include("server:libs:modules:components:accelo")
 include("server:libs:modules:components:active-campaign")
 include("server:libs:modules:components:affinity")
+include("server:libs:modules:components:ai:ai-text-analysis")
 include("server:libs:modules:components:airtable")
 include("server:libs:modules:components:aitable")
 include("server:libs:modules:components:app-event")
@@ -268,7 +269,6 @@ include("server:libs:modules:components:jotform")
 include("server:libs:modules:components:json-file")
 include("server:libs:modules:components:json-helper")
 include("server:libs:modules:components:keap")
-include("server:libs:modules:components:llm:ai-text-analysis")
 include("server:libs:modules:components:llm:amazon-bedrock")
 include("server:libs:modules:components:llm:anthropic")
 include("server:libs:modules:components:llm:azure-openai")
